requirements.txt
llama-index-core
llama-index
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-llms-groq
llama-index-llms-gemini
llama-index-utils-workflow
llama-parse
python-dotenv
streamlit
torch
llama-index-llms-deepseek